+++ tests/PhpWord/Tests/TOCTest.php
@@ -78,5 +78,9 @@ class TOCTest extends \PHPUnit_Framework_TestCase
             $this->assertEquals($depth, $savedTitles[$i]['depth']);
             $i++;
         }
+
+        TOC::reset();
+        $this->assertEquals(0, count(TOC::getTitles()));
+
     }
 }